Move all the specs to test folder
