+++ test/endsWith.spec.ts
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 import RulesEngine from "../src/index";
-import { RuleSet } from "../types/types/rules";
+import { RuleSet } from "../src/types/schema";
 
 const RULE_DEFINITION: RuleSet = {
   version: 1,
