+++ test/utils/isNumber.spec.ts
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { isNumber } from "./isNumber";
+import { isNumber } from "../../src/utils/isNumber";
 
 describe("isNumber", () => {
   it("returns true for a given number", () => {
