+++ test/utils/isUndefined.spec.ts
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { isUndefined } from "./isUndefined";
+import { isUndefined } from "../../src/utils/isUndefined";
 
 describe("isUndefined", () => {
   it("returns false when passed a string", () => {
